move MockitoSugar class to core module
