+++ mist/master/src/test/scala/io/hydrosphere/mist/master/interfaces/http/DevApiSpec.scala
@@ -2,13 +2,13 @@ package io.hydrosphere.mist.master.interfaces.http
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import io.hydrosphere.mist.core.Action
+import io.hydrosphere.mist.core.{Action, MockitoSugar}
 import io.hydrosphere.mist.core.CoreData._
-import io.hydrosphere.mist.master.JobDetails.{Status, Source}
-import io.hydrosphere.mist.master.{MockitoSugar, ExecutionInfo, MasterService}
+import io.hydrosphere.mist.master.JobDetails.{Source, Status}
+import io.hydrosphere.mist.master.{ExecutionInfo, MasterService}
 import io.hydrosphere.mist.master.interfaces.JsonCodecs
 import io.hydrosphere.mist.master.models.{DevJobStartRequest, DevJobStartRequestModel}
-import org.scalatest.{Matchers, FunSpec}
+import org.scalatest.{FunSpec, Matchers}
 
 import scala.concurrent.Promise
